Add Sirius gateway mock back in (useful for testing)
We used to have the Sirius gateway mock turned on by default
locally. When we changed our approach to testing statuses, I took
it out. However, it can be useful for testing scenarios which are
tricky to test otherwise; for example, caching.

Add the mock back in but as a manual testing tool, with
instructions on how to enable it.