LPAL-765 Add the AWSManagedRulesCommonRuleSet to our WAF config (#1060)
* Add the AWSManagedRulesCommonRuleSet to our WAF config

This is to bring us in line with other teams and webops CoP
recommendations.

This ruleset blocks requests which match a variety of common
attack patterns, such as requesting potentially vulnerable
file paths, sending cross-site scripting payloads in request bodies
and similar.

* Set action for requests matching AWSManagedRulesCommonRuleSet to count

This allows us to turn the ruleset on for our WAF in prod without
blocking requests. After monitoring for a while, we can then decide
whether we can switch on blocking instead.