Rationalise service-front Dockerfile to enable installing with platform reqs
* Remove the separate composer image previously used to install
deps. Instead, just install composer into the main image (and
remove it after using it). This reduces the amount of stuff we
have to download, and also means we are calling composer in
an environment with all the required dependencies in place.
* Remove the --ignore-platform-reqs flag from composer call. This
forces composer to only install packages for which we have required
extensions turned on, the right PHP version etc.
* Add a --no-cache flag to composer call. This prevents composer
from keeping a local copy of downloaded packages outside the
/app/vendor directory.