Properly migrate to Cypress 10
On looking at this, I realised that there were big API changes
between Cypress 9 and 10 which we hadn't fully accounted for.

Update our tests to fit the migration requirements, and also
move to the @badeball version of cypress-cucumber-preprocessor
(the official community package for this).

One of the main changes is the requirement to be explicit about
loading some files (e.g. e2e.js support file) and some files
no longer being autoloaded (e.g. plugins/index.js). The mechanism
for configuring plugins has also changed, so that had to be rewritten.