Streamline setting env vars for cypress
CYPRESS_* environment variables and variables passed with
-e name=value have the same effect in Cypress: both create
variables within the Cypress environment (NOT in the OS environment)
which are scoped to cypress.

Given this is the case, remove the distinction between -e vars and
CYPRESS_* vars, as they have the same effect.

Add in additional -e vars which specify the location of stepDefinitions
and filterSpecs=true. The latter is required for the latest
version of the cypress-cucumber-preprocessor to work: without this,
cucumber tags on Cypress features like "@Signup" are ignored. With
this setting at true, the new preprocessor filters by cucumber tag,
enabling us to retain our current approach, where we run the @Signup
features followed by others in a specific order.

Note that the cypress-tags script which used to do the filtering
has now been folded into the cucumber preprocessor, so we just
call cypress now.