Sometimes responses are JSON, sometimes HTML...
The same POST request to the same endpoint will return JSON
if there are no errors, otherwise HTML.

To account for this, modify the moj.Helpers.ajax() function
to check the response type and parse it to JSON if its
content-type is application/json.

This is terrible (the POST should always receive JSON *or* HTML),
but its a lot of work to unpick. jQuery used to manage this
automagically for us, but we have to do it ourselves now.