LPAL-846 Remove unused libraries and ensure installed libraries meet PHP version and platform reqs (#1074)
* Update service-front composer deps

* Rationalise service-front Dockerfile to enable installing with platform reqs

* Remove the separate composer image previously used to install
deps. Instead, just install composer into the main image (and
remove it after using it). This reduces the amount of stuff we
have to download, and also means we are calling composer in
an environment with all the required dependencies in place.
* Remove the --ignore-platform-reqs flag from composer call. This
forces composer to only install packages for which we have required
extensions turned on, the right PHP version etc.
* Add a --no-cache flag to composer call. This prevents composer
from keeping a local copy of downloaded packages outside the
/app/vendor directory.

* Update service-api composer.lock

* Upgrade laminas-mvc-i18n in service-api and downgrade accepted stability

This brings service-api in line with service-front, both in terms
of the accepted package stability and the version of laminas-mvc-i18n.

* Rationalise service-api Dockerfile to enable installing with platform reqs

* Remove composer container and install composer into main container.
Remove it when used. Less fuss, less download.
* Add explicit intl PHP dep, so we can install with platform reqs.
* Remove --ignore-platform-reqs flag on composer call.

* Rationalise service-admin Dockerfile to enable installing with platform reqs

* Remove extra composer container
* Remove --ignore-platform-reqs flag on composer call

* Update service-admin composer deps

* Fix psalm errors for service-front

* Fix psalm lint errors for service-admin

* Remove unused composer deps from service-front

* Prep for unused composer deps scanning in service-api

* Remove unused composer dependencies from service-api

* Prep for scanning unused composer deps in service-admin

* Remove unused composer deps from service-admin

* Remove unused scanner config

This is something we won't be doing regularly (we are more careful
about dependencies than our predecessors), so we don't really need
to spread this config over our whole code base.

It's relatively trivial to incorporate again; see
https://github.com/Insolita/unused-scanner for instructions.

* Add erroneously-removed laminas-cli package back in

* Add laminas-router back into service-admin composer deps

This was previously being pulled in by the laminas-mvc dep,
but we don't need the whole of laminas-mvc for service-admin
to work correctly.

Just pull in the router component we need.