Fix local builds
Local composer install was failing inexplicably, so update
the Makefile to fix, and configure service-admin to allow
laminas composer plugins.

This is likely to be due to composer 2.4 being more strict, and
possibly changing how composer is called from the docker image.

Update to composer:latest locally, same as we're using in our
Dockerfiles.