LPAL-888 cypress-cucumber-preprocessor (and associated cypress) upgrade (#1077)
* Properly migrate to Cypress 10

On looking at this, I realised that there were big API changes
between Cypress 9 and 10 which we hadn't fully accounted for.

Update our tests to fit the migration requirements, and also
move to the @badeball version of cypress-cucumber-preprocessor
(the official community package for this).

One of the main changes is the requirement to be explicit about
loading some files (e.g. e2e.js support file) and some files
no longer being autoloaded (e.g. plugins/index.js). The mechanism
for configuring plugins has also changed, so that had to be rewritten.

* Update to cypress 10.5.0

* Streamline setting env vars for cypress

CYPRESS_* environment variables and variables passed with
-e name=value have the same effect in Cypress: both create
variables within the Cypress environment (NOT in the OS environment)
which are scoped to cypress.

Given this is the case, remove the distinction between -e vars and
CYPRESS_* vars, as they have the same effect.

Add in additional -e vars which specify the location of stepDefinitions
and filterSpecs=true. The latter is required for the latest
version of the cypress-cucumber-preprocessor to work: without this,
cucumber tags on Cypress features like "@Signup" are ignored. With
this setting at true, the new preprocessor filters by cucumber tag,
enabling us to retain our current approach, where we run the @Signup
features followed by others in a specific order.

Note that the cypress-tags script which used to do the filtering
has now been folded into the cucumber preprocessor, so we just
call cypress now.

* Wait for the cypress-cucumber-preprocessor config to be applied

As this config is applied asynchronously, we have to wait for
it to be applied (with the await keyword). If we don't the
filterSpecs setting is ignored and features aren't filtered
(tricky one to figure out).

* Downgrade to cypress 10.3.1

* Tell local cypress UI where step definitions are

* Variable names must match those recognised in our cypress step definitions

* Remove hard-coded baseUrl for cypress test suite

* CYPRESS_baseUrl and -e baseUrl don't appear to be equivalent

Setting the baseUrl with the -e flag to cypress doesn't appear
to work correctly.

Go back to setting it via CYPRESS_baseUrl in the environment.
Same for adminUrl.

* Makefile readability improvements

* Clean up cypress config file

* Make lines a bit shorter
* Move variables based off userNumber closer to its declaration
* Remove setting default adminUrl

Now that we are explicitly passing adminUrl and baseUrl into
our scripts, removing this fallback should more quickly cause
errors if they are not set.

* Upgrade cypress to 10.7.0

* Additional explanation for CYPRESS_RUNNER_* vars in Makefile