LPAL-769 Session fixes based on ASVS recommendations (#1069)
* Remove session ini settings which have no effect

hash_function and hash_bits_per_character are deprecated in PHP 8.

* Set SameSite=Strict on session ID cookie

* Clear user session cookie when their session has expired

If the user has no time left in their session, clear their session
cookie. This is an ASVS requirement that cookies relating to
invalid sessions are removed from the browser, regardless of
whether those cookies would make it possible to access resources
or not.

* Fix AuthController unit test

We now explicitly destroy the session when it expires, to
ensure we remove the cookie and any stored data relating to it.

* Cannot set SameSite on session cookie

This prevents the redirect to govuk pay from working.