LPAL-957 jQuery upgrade prep (#1082)
* Add configuration for eslint

* Fix typo

* Remove redundant JS module

The functionality this used to provide (show the help message
when the user selects the fee they want to pay) seems to have
been taken over by another module. This code was left here
however, but it appears to have no purpose, so remove it.

* Remove unused jQuery plugin

* Remove unused session popup JS file

* Clean up moj.dashboard.js and remove unused code

The changeMobileActions function uses a selector which looks for
"tr" (table row) elements on the page. As we don't use any of these,
this isn't doing anything, so remove it.

* Fix CSS for dashboard search button

* Ensure that the searchbox is styled conditionally (only on dashboard)

* Use full name for GOVUK components in Twig templates

* Wait for popup opener link to be ready before clicking

* Remove unnecessary RUN step from Dockerfile

* Upgrade grunt-contrib-uglify

This enables uglify to cope with the const keyword, which the
version we were using is unable to deal with.

* Use same packages locally as in CI for node-build-assets

When building JS and CSS assets locally, use the same packages
as are used in the CI build, by installing from package.json.

Previously, we were installing dependencies manually in the
Dockerfile, then again in start.sh.

* If paths are set correctly, no need to download sass separately

* Use npm ci instead of npm install

npm install updates dependencies, while npm ci does not.

* Remove unnecessary comment

* Remove double arrow when <details> is polyfilled

* Remove unused JavaScript

* Remove sticky-nav JS from Grunt build

* Remove unused Who Are You? client-side JavaScript

* Remove unused has-scrollbar jQuery plugin from Gruntfile

* Install eslint in dev

* Add cypress tests for client-side validation in person form

* Name in form duplicates a name elsewhere in the application
(attorney duplicates name of donor in this case)
* Age of person in form < 18
* Age of person in form > 100

* Install eslint dependencies for standard JS style