Include twig templates in coverage report
This makes it possible to see how much of our twig templates
is covered in unit tests.

The trick is to look for block_<block name> functions in the
coverage report for a particular twig template. For example,
as we test the donorGuidance block in the date check twig
template, we can see the test coverage for that by looking
at the block_donorGuidance() method in the coverage report.

Unfortunately, it's not possible to get nice clean names for
the twig templates. This makes it difficult to locate the right
part of the coverage report to see how the coverage looks for
the compiled twig template.

Also note that you will have to run the tests twice to get coverage
for the twig templates: once to compile the templates, then again
so that phpunit can load the compiled templates and instrument
them for coverage measurement.