Modify criteria for when continuation sheet 2 is generated
Use more naturalistic criteria for when cs2 is generated, which
map more readily to manual test cases. This makes it simpler
to determine whether the test cases are covered thoroughly, and
makes it easier to marry up the automated and manual cases (even
though there is some redundancy here).

Also note that some of the primaryAttorneyDecisions->how values
only occur when there are multiple attorneys. However, we still
count the attorneys for the purpose above, making it easier
to work things out when manual tests fail.