Date check/continuation sheet feature (#1100)
* Revert "Revert "Merge pull request #983 from ministryofjustice/LPAL-705-prompt-signing-continuation-sheets" (#1041)"

This reverts commit 68c9885857fed1fe55bad10be4050b70bf244cb9.

* Amend heading for when donor can't sign [LPAL-875]

* Refactor c sheet logic into separate service

* Remove unused var dateError

* Move service to CS ViewModel Helper

* Move applicants logic to helper

* Setup to test main panel-border

* Add expectations and test cases up to c. sheet 2

* Render entire date check form in view test wip

* Add test cases for donor and attorney

* Refactor and add comments

* Fix CompleteController c.sheets section

I removed the inheritence relationship between AbstractService and
ContinuationSheet service some commits ago, which broke the
CompleteController. This means the service class can be directly
instantiated in the controller without the need to go through the
standard instantiation process via the ControllerAbstractFactory

* Remove unneeded tests for CompleteController

* Clean up controller tests

* Fix comma type for Cypress test

* Changes from PR comments

* Change content for PF LPAs with c. sheet 3

This updates content for where the donor can't sign (cs3) in combination
of other c sheets being generated

* Change content for HW LPAs with c. sheet 3

* Update cypress tests to reflect content change

* Amend applicant text in date check tool

This applies when applicant is the donor, and the donor cannot sign.

Cypress tests are extended to check this.

* Amend form error messages when donor cannot sign (LPAL-890)

* Fix mistaken cypress test assumption

The applicant can be either the donor or the attorney according to the
UI. The title and errors all normally refer to them as 'the applicant'
unless the applicant is the donor and the donor can't sign, then it says
'the person signing on behalf of the donor'.

* Fix import in cypress support file

* Wrap donor name in check signing dates page

Where the donor has a long name and is unable to sign, the donor's
name can overlap the date boxes on the check signing dates page.
Set a maximum width to prevent this from happening.

* Clean up syntax in DateCheck files

* LPAL-890 Add unit tests for DateCheck when donor can't sign

* LPAL-890 Implement code to change date check errors when donor cannot sign

* LPAL-875 Test text shown for person signing for donor

* LPAL-875 Implement additional text for person signing for donor

Where a person is signing on behalf of the donor, the check date
page should explain that continuation sheet 3 is signed both
by that person and two witnesses.

* Fix broken cypress test

Broken because the text changed to include ", followed by two witnesses".

* Reword and fix tests for signing on behalf of donor/applicant

Relates to LPAL-890 and LPAL-891

* Fix missed 'a person signing on behalf' text

Agreed to use "the person signing on behalf", but missed a
couple of places where this is used.

* Remove references to section 9 when continuation sheet 3 is present

* Include twig templates in coverage report

This makes it possible to see how much of our twig templates
is covered in unit tests.

The trick is to look for block_<block name> functions in the
coverage report for a particular twig template. For example,
as we test the donorGuidance block in the date check twig
template, we can see the test coverage for that by looking
at the block_donorGuidance() method in the coverage report.

Unfortunately, it's not possible to get nice clean names for
the twig templates. This makes it difficult to locate the right
part of the coverage report to see how the coverage looks for
the compiled twig template.

Also note that you will have to run the tests twice to get coverage
for the twig templates: once to compile the templates, then again
so that phpunit can load the compiled templates and instrument
them for coverage measurement.

* Fix message where person signing for applicant in future

Where the donor is the applicant, and the donor (and therefore
applicant) can't sign for themselves, and a future date is
entered in the date checker for the signature date of the person
signing on behalf of the applicant, show appropriate error
message.

* Fix message when person signing for applicant before attorneys

When the person signing on behalf of the applicant, who is also
the donor, signs before an attorney, show appropriate error message.

* Add cypress tests for person signing on behalf of applicant

Cases covered as per LPAL-891:

* Person signing on behalf of applicant signature date in future
* Person signing on behalf of applicant signature date before
attorney signature

* Add note to fixture SQL that donor is applicant for application

* Don't pass unused variables to view

* Amend donor date check prompt for visually-impaired users

The visually-hidden text should reflect whether the area of the
date check is for the donor or the person signing on their behalf,
same as the visible text.

* Modify criteria for when continuation sheet 2 is generated

Use more naturalistic criteria for when cs2 is generated, which
map more readily to manual test cases. This makes it simpler
to determine whether the test cases are covered thoroughly, and
makes it easier to marry up the automated and manual cases (even
though there is some redundancy here).

Also note that some of the primaryAttorneyDecisions->how values
only occur when there are multiple attorneys. However, we still
count the attorneys for the purpose above, making it easier
to work things out when manual tests fail.

* Convert conditional into switch/case for better coverage reports

* Fix faulty continuation sheet logic and complete unit tests

* Fix condition which produces a cs2

The condition was triggered by multiple replacement attorneys
acting jointly, but should be jointly and severally. This was an
error introduced when transcribing from the acceptance criteria
into the code.

* Adjust visually-hidden text for donor/person signing for donor

* Additional seeding dependency conditions in docker-compose.yml

* Seeding relies on dynamodb not localstack

* Make Cypress tests for date check page more generic

Pass names to be looked for in the test as parameters so the
support functions are more reusable.

Co-authored-by: Mabel Li <mabel.li@digital.justice.gov.uk>