Make session timeout cypress test less precise
The session timeout cypress test sets the session timeout to 302
seconds, so that the popup can be forced to open in the test.
However, the response being returned when the timeout is updated
states that the session length is 301 seconds. This suggests there's
some lag between the session length set by the request, and the
amount of time left when the response is returned.

To fix this, change the expectation that the response states 302
seconds as the remaining session time, to expect *at most* 302
seconds instead.