Modify config to enable accessing different paths on Sirius gateway
The setting for the processing status URL includes the path to the
specific LPA tool resource, i.e. the status of a single LPA accessible
by its A-ref.

As we need to use this URL in another place, to healthcheck the
Sirius gateway at /healthcheck, modify the configuration so that
the URL only includes the scheme, domain and version part of the path.
We can then append /lpa-online-tool/lpas/<LPA ID> or /healthcheck to
that URI as necessary.