LPAL-755 Upgrade service-pdf to PHP 8.1 (#997)
* Update to PHP 8.1 and opg-lpa-datamodels 14

* Use the image renderer to create barcodes

* composer require vimeo/psalm

* Fix psalm issues and remove unused code

* Clean up method signatures to match superclasses'

* Support rendering barcodes as either PDF or PNG

* Use image rendering for barcodes

* Simplify writing PNG barcode to file

After re-reading the PHP docs, I realised it is possible to
optionally pass a filename to imagepng() to write the PNG straight
to a file.

* Remove unnecessary extra PDF classes

* Refactor SqsWorker to use separate PdfRenderer

By isolating the PDF rendering from the worker, we can more
easily unit test it, and more easily generate PDFs to order
for the purposes of performing a comparison (e.g. if we change
how PDFs are rendered).

* Remove some deprecated and redundant files

* Modify local PDF render script to use PdfRenderer

Note that this doesn't work locally at the moment and only
functions properly when run from inside the lpa-pdf-app docker
container.

* Refactor to enable test-specific config for PDF generation

This enables paths to be specified for use during local runs,
which makes the destination of generated PDFs predictable
and reportable. In turn, this enables us to get hold of the PDFs
for the purpose of comparing how they look vs. previous renders.

* Reorganise writing PDF files to disk

Intermediate PDF files are removed from disk as soon as the PDF
object is done with. This makes refactoring difficult, as you
lose the PDF file as soon as its object is garbage collected.

To manage this, modify the PDF renderer so that it returns both
the filepath of the PDF and its content (as a binary string).
This can then be written to the local filesystem (for generating
PDFs for testing purposes) or to S3.

The intermediate PDFs are still garbage-collected, and in the
online worker will be removed as soon as they are put to S3.
However, we can manually write the PDF content to the local disk
for testing as we don't lose the PDF content immediately.

* Don't pass Config around to PDF classes

* No longer need to pass Config to PDF objects

* Make CLI PDF generator more useful

* Get the name of the LPA JSON file to process from the command line
arguments.
* Include the LPA ID in the output filename.

* Scale size and offsets for PNG barcode

Generate a PNG barcode which is A4-sized, modifying the offsets
and factor (which scales the barcode bar widths and height) so that
the barcode is in the correct position. This will still need
to be converted to a PDF anyway, as Pdftk can't cope with using
a PNG as a stamp file and fails silently.

* Use TCPDF to convert PNG barcode to PDF

* Remove ZendPDF dep, upgrade other packages to PHP 8.1 versions

* Bring Dockerfile in line with others (LPAL-846)

* Remove composer container
* Run composer inside main PHP container without --ignore-platform-reqs
* Make the build dependencies directory align with how we do things in other
Dockerfiles.
* Move some build deps which were accidentally kept in the image so
they are removed after build completes.
* Remove the slightly-confusing swapping round of WORKDIR and just
save files into the right directory.

* Minor syntax cleanup for script

* Fix to composer 2.4.1

This brings service-pdf in line with other components.

* Add missing type hint

* Make it clear that these files are for testing

* Add explanation of the TODO