Make session timeout checks more lenient
When setting the session timeout to 0, we occasionally get a
response where the session time remaining is 1. This causes
the tests to fail.

Make the comparison more lenient so that we allow for a range of
integer values to be returned, ranging from the value we tried to
set to 5 seconds around than that value.