Update and pin to version 14.0.4 of opg-lpa-datamodels
To reduce the risk of releasing new versions of opg-lpa-datamodels,
pin the version to whatever works with the branch.

In future, this shouldn't be necessary, as we plan to move
the opg-lpa-datamodels repo into the main opg-lpa repo, at which
point versioning and tagging will no longer be an issue.