LPAL-1033 Fix trust corporation error messages (#1161)
* Add a guard to prevent redundant error messages on company number

This copies the technique of LPAL-325, which ignores any error
messages set to ''. However, a better solution would be to prevent
these error messages in the first place: having both "not blank"
and "at least one character long" is redundant, and should be
fixed in the data model itself.

* Remove references to trust corp errors which no longer occur

Changes to opg-lpa-datamodels mean that errors about the minimum
length of trust corporation name and number will no longer occur,
so remove the error message mapping for the appropriate keys.

* Add cypress helper to quickly fill out a form

* Extend attorney cypress test to check name and number errors

Ensure that the system error messages mentioned in LPAL-325 and
LPAL-1033 are not visible when trust corporation name and number
are blank.

* Update and pin to version 14.0.4 of opg-lpa-datamodels

To reduce the risk of releasing new versions of opg-lpa-datamodels,
pin the version to whatever works with the branch.

In future, this shouldn't be necessary, as we plan to move
the opg-lpa-datamodels repo into the main opg-lpa repo, at which
point versioning and tagging will no longer be an issue.