LPAL-1006 Better Sirius gateway healthcheck (#1157)
* Expose different localstack and postgres ports

Use different local ports for these services so they don't
clash with other services using them (e.g. Sirius).

* Call the /healthcheck endpoint to see if Sirius gateway is up

* Fix unit tests for PingController

* Enable integrating with local opg-data-lpa

If you have an opg-data-lpa instance running locally on localhost,
this will use its /healthcheck endpoint to check the Sirius gateway
status.

This allows us to see correct output from our /ping endpoint,
including the Sirius gateway status.

* Modify config to enable accessing different paths on Sirius gateway

The setting for the processing status URL includes the path to the
specific LPA tool resource, i.e. the status of a single LPA accessible
by its A-ref.

As we need to use this URL in another place, to healthcheck the
Sirius gateway at /healthcheck, modify the configuration so that
the URL only includes the scheme, domain and version part of the path.
We can then append /lpa-online-tool/lpas/<LPA ID> or /healthcheck to
that URI as necessary.

* Include gateway URI in PingController error logging

So we can see what's happening when this is deployed in the dev
environment.

Also changed to info() log level so it shows up in the logs.

* Send content-type header in case opg-data-lpa wants it

Also fix some inconsistent snake case variable names.

* Permissions for api to access opg-data-lpa /healthcheck path

* Remove wildcard from sirius_api_healthcheck_arn

Co-authored-by: Sam Wilkinson <sam.wilkinson@digital.justice.gov.uk>