Checkout page requires you to have just visited the fee reduction page
If this is not the case, the checkout page incorrectly assumes
the LPA is incomplete because you haven't selected a certificate
provider.

This is frankly a series of daft design choices: to determine whether the
LPA has been completed and is ready to be confirmed and finished,
check which page we were just on. Then, to make matters worse,
if we haven't come from the right page, assume the LPA is
incomplete because the certificate provider was skipped. In effect,
the completeness of the LPA is conditional on which part of the UI
you just visited.