Slim down DateCheck.feature
Rather than going to the dashboard and clicking on links to
reach the date check tool, go direct to the page.

This also removes the reliance in the tests on the LPA appearing
on the first page of the dashboard. By either searching for the LPA
(to check the link from the dashboard to the date check page),
or bypassing the dashboard, it doesn't matter how many LPAs are
present on the dashboard.

(What I found was that running the tests a lot manually added so
many LPAs that the links expected by the date check tests
have fallen off the bottom of the page.)