LPAL-797 Mend the cypress tests so they can all run standalone (#1187)
* Provide better feedback when Python API client fails

* Move bulk of sign-up related tests out of Signup.feature

Do the minimum possible number of assertions in this feature,
as it's run before multiple other features.

Move tests specific to signup and changing details into
the feature dedicated to that.

* Start repairs for CheckoutPFClone

* Modify the -w option so it's possible to set a non-donor registrant.

* Decouple setting repeat application from LPA type in tests

* Enable setting payment independently of LPA type

* Checkout page requires you to have just visited the fee reduction page

If this is not the case, the checkout page incorrectly assumes
the LPA is incomplete because you haven't selected a certificate
provider.

This is frankly a series of daft design choices: to determine whether the
LPA has been completed and is ready to be confirmed and finished,
check which page we were just on. Then, to make matters worse,
if we haven't come from the right page, assume the LPA is
incomplete because the certificate provider was skipped. In effect,
the completeness of the LPA is conditional on which part of the UI
you just visited.

* Enable adding attorney as correspondent

* Separate password change tests from password reset

Also put the reset tests into their own area with signup
occurring beforehand, so that they can run standlone.

* HW fixture for CheckoutHW.feature should not be repeat application

* Improve cookie feature (cypress tests)

* Slim down DateCheck.feature

Rather than going to the dashboard and clicking on links to
reach the date check tool, go direct to the page.

This also removes the reliance in the tests on the LPA appearing
on the first page of the dashboard. By either searching for the LPA
(to check the link from the dashboard to the date check page),
or bypassing the dashboard, it doesn't matter how many LPAs are
present on the dashboard.

(What I found was that running the tests a lot manually added so
many LPAs that the links expected by the date check tests
have fallen off the bottom of the page.)

* Remove all uses of "quickly fill out"

* Go direct to view LPA page rather than via link

This is to deal with the issue where the LPA may be on page 2
of the dashboard.

* Add a natural language expression for visiting date check pages

* Fix bug where cypress doesn't type the full string I asked it to...