LPAL-1043 Allow HTML characters and trailing/leading spaces in passwords (#1195)
* Disable StripTags and StringTrim for password fields

We apply StripTags and StringTrim to all fields in all forms.
This means that characters like <> and trailing and leading spaces
are removed from passwords, resulting in misleading error messages.

Don't apply those filters to fields in the new account, sign in
and change password forms.

Note that we are protected against SQL injection by the fact
that we use parameterised queries for all SQL operations (via the
Laminas framework). We don't construct SQL queries manually
ourselves by joining strings together from user input. So while it
is generally best to remove HTML tags where possible, we are
well protected when loosening this for a handful of password fields.

Also note that the user's password is never rendered to screen. This
makes any attempt to inject HTML into a password pointless, as we
never display it back to the user once it has been saved.

* Add tests for changing passwords to contain HTML and trailing/leading spaces

* Add tests for registering with password containing HTML and spaces