pdftk check can't be done in constructor
The check for whether pdftk is present can't be done in the
constructor, as the php-fpm process appears to be unable to
detect the presence of the script at that point.

By moving it out into the getPdftkCommand() method, and caching
the results of the check (not sure if this has any effect given
how the PDF poller works, but can't hurt), we are able to apply
the test for the presence of the script correctly in the real
container.