LPAL-1105 Return 404 when trying to download a PDF for an incomplete LPA (#1264)
* Return a 404 if an LPA's PDFs are not available

The old code returned a ViewModel, which was then supposed to
be rendered in the Download Twig index template. However, this
template doesn't exist, which is why we were seeing 500 errors.

Fix this by using the notFoundAction(), which returns a templated
404 response.

* Fix local bug where nginx is returning a 403 response

nginx does not return valid HTML5 if the directive

return 403;

is used for a location. To work around this, only test the
doctype name if the document has a doctype.

* Add test that downloading unready PDF returns a 404

* Update phpstorm config

I had to do this because I wanted to be able to debug unit test
failures. I really don't like that this is checked in, as it
will be different for every user of the project.

* Fix unit tests broken by calling notFoundAction() in DownloadController

Unit tests for controllers are not straightforward, but this
at least makes it possible to fix the unit test broken by returning
a 404 when a non-existent PDF download is attempted.

Co-authored-by: Elliot Smith <elliot.smith@justice.digital.gov.uk>