Use standard array access by keys
Note that one of the accesses into the Payment object looked like this:

$payment->state->code

Use of magic getter methods like this is deprecated in PHP 8.2,
and can't be mocked in a unit test easily because the state
object is stdclass and not an array.

To get around this, I added a getStateCode() method which can be
mocked in tests, which just accesses $payment->state->code internally
in the Payment class.