LPAL-1114/LPAL-1142 Add X-Ray tracing (with sampling) to service-api (#1320)
* Add opentelemetry package

Note that this requires the laminas-diactoros package, which
provides an up to date psr7 implementation.

If php-notifications-client is ever updated to support guzzlehttp
v2.4+, we can remove the diactoros dependency.

* POC of adding opentelemetry to service-api

* Encapsulate opentelemetry in our own class

* Add tracing to part of the API application

* Export spans to the standard Laminas log

* Move telemetry classes to shared

* Update dependencies required for opentelemetry in shared

* Use X-Amz-Trace-Id header with OpenTelemetry

Extract necessary elements from the non-standard X-Amz-Trace-Id
header to populate the standards-based OpenTelemetry objects.

This is currently hard-coded to a single value in the correct
AWS trace ID format, which is all we need in dev to confirm
we can work correctly with Amazon-formatted trace IDs.

* Clean up and psalm check

* Move Telemetry into a module

This should mean we can move away from a singleton pattern,
making it easier to configure the tracer from our config file,
and use events to manage spans around application calls.

* Create telemetry tracer in standard Laminas factory fashion

* Use the opentelemetry AWS library to propagate/extract headers

AWS X-Ray headers are in a slightly different format from
those required by opentelemetry. The AWS otel library deals with
extracting and propagating the X-Ray header so it's usable by otel.

* Enable addition of attributes to otel spans

* Configure service-api to send telemetry to local collector

* Use URL config for telemetry exporter instead of host + port

* Add guard so events are only triggered if event manager is set

* Add protobuf PHP extension

This is an optional opentelemetry dependency, but makes the
PHP library more performant: see
https://github.com/opentelemetry-php/opentelemetry-php#optional-dependencies

* Log telemetry URL to help debug dev env

* Use localhost IP address rather than domain name

* Fix comment now telemetry config has changed

* Revert to our own code for generating trace and span IDs

The AWS contrib fails to accept the x-amz-trace-id header if it
doesn't contain a Parent key.

Rather than do that, generate our own Parent keys where they are
missing, with a fresh span ID.

* Add note about correct format for aws-otel-collector URL

* Use correct return type when starting spans

* Documentation improvements

* Fix module name for SpanInterface

* Use UDP transport to export X-Ray traces

* Export using UDP to port 2000 on aws-otel-collector

* Rework AWS X-Ray tracing

* Remove opentelemetry dependency

* Use UDP on port 2000 for reporting telemetry

* Segment stop events for telemetry no longer need a name

The event causes whatever the currently-active segment is to
be stopped. It therefore no longer needs a segment name.

* Put Telemetry exporters in their own directory

* Only start X-Ray tracing if x-amz-trace-id header has 'Root' key

* Add method to get the ID of the current segment

* Improve and correct telemetry module docs

* Add unit tests for MakeShared\Telemetry\Tracer

Also pulled in Mockery as a dependency.

* Add unit tests for MakeShared\Telemetry\Segment

* Add unit tests for MakeShared\Telemetry\TelemetryEventManager

* Add unit tests for MakeShared\Telemetry\Module

* Exercise the console exporter

* Remove unused deps for shared/

* Use SimpleLoggerTrait for consistency

* protobuf extension is no longer needed

This improves performance when using opentelemetry-php, but we
no longer are, so remove it.

* Add tests for 'sampled' implementation

* Implement sampling for X-Ray

This is on the exporter, so we just ditch Segments which have
their sampled property set to false.

* Remove the indirection and unintuitiveness of shouldBeExported()

* Telemetry module has to be correctly namespaced to load

* Log any errors when exporting telemetry

* Remove laminas-diactoros

This was required for opentelemetry, which I've removed.

* Fix docs

---------

Co-authored-by: Elliot Smith <elliot.smith@justice.digital.gov.uk>